PISTON-1143: Reconciles the voicemail schema documentation.
At some point, the `voicemail` document was renamed as `vmboxes`, but the related documentation files were not  renamed to match. As a result, a hand-edited `voicemail.md` document diverged considerably from the auto-generated `ref/vmboxes.md` document. The contents of the `voicemail.md` have been reordered to reduce the degree of difference from the auto-generated file, and some minor corrections were made to the text.