KTEL-10: Replace colons by dashes on fax attachments names (#6734)
* KTEL-10: Replace colons by dashes on fax attachments names

When teletype sends a fax attachment, its name includes a time format
that includes colons. Colons are not allowed characters for filenames in
Windows and so many SMTP relays and mail servers will auto-correct the
colon to a dash or underscore character. However, if the end user's
server does not correct this, the resulting attachment is unopenable on
Windows.

* KTEL-10: Split nested functions